Factored out all util code into new mdp module
